data-ingestion/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.4
azure-search-documents==11.4.0
azure-core==1.29.5
python-dotenv==1.0.0
schedule==1.2.0
